chore: add minor improvements
