+++ pubspec.yaml
@@ -19,4 +19,5 @@ dependencies:
   meta: ^1.9.1
 
 dev_dependencies:
+  collection: ^1.18.0
   flutter_lints: ^3.0.1
